requirements.txt
# Content Workflow Automation Dependencies

# Transcription
openai-whisper>=20231117

# YouTube API
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Optional: for video processing
# ffmpeg-python>=0.2.0
# moviepy>=1.0.3
